+++ NEWS.md
@@ -14,6 +14,7 @@
   coverage bin must reach to be reported as covered.
 - Added a warning for potential infinite loops in processes without
   sensitivity and lacking any wait statements (from @NikLeberg).
+- Fixed a crash when `'last_value` is used with record types (#1043).
 
 ## Version 1.14.1 - 2024-10-26
 - Fixed an error when using the `work` library alias and the working